refactor: extract storage standard impl
