+++ contract/src/core.rs
@@ -1,9 +1,6 @@
-use near_contract_standards::{
-    fungible_token::{core::FungibleTokenCore, resolver::FungibleTokenResolver, Balance},
-    storage_management::{StorageBalance, StorageBalanceBounds, StorageManagement},
-};
+use near_contract_standards::fungible_token::{core::FungibleTokenCore, resolver::FungibleTokenResolver, Balance};
 use near_plugins::{pause, Pausable};
-use near_sdk::{env, json_types::U128, near, AccountId, NearToken, PromiseOrValue};
+use near_sdk::{env, json_types::U128, near, AccountId, PromiseOrValue};
 
 use crate::{Contract, ContractExt};
 
@@ -37,34 +34,6 @@ impl FungibleTokenCore for Contract {
 
     fn ft_balance_of(&self, account_id: AccountId) -> U128 {
         self.token.ft_balance_of(account_id)
-    }
-}
-
-#[near]
-impl StorageManagement for Contract {
-    #[payable]
-    fn storage_deposit(&mut self, account_id: Option<AccountId>, registration_only: Option<bool>) -> StorageBalance {
-        self.token.storage_deposit(account_id, registration_only)
-    }
-
-    #[payable]
-    fn storage_withdraw(&mut self, amount: Option<NearToken>) -> StorageBalance {
-        self.token.storage_withdraw(amount)
-    }
-
-    #[payable]
-    fn storage_unregister(&mut self, force: Option<bool>) -> bool {
-        self.assert_not_in_denylist(vec![&env::predecessor_account_id()]);
-
-        self.token.internal_storage_unregister(force).is_some()
-    }
-
-    fn storage_balance_bounds(&self) -> StorageBalanceBounds {
-        self.token.storage_balance_bounds()
-    }
-
-    fn storage_balance_of(&self, account_id: AccountId) -> Option<StorageBalance> {
-        self.token.storage_balance_of(account_id)
     }
 }
